pyproject.toml
[project]
name = "sis-course-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "selenium>=4.40.0",
    "webdriver-manager>=4.0.2",
]
